Restructred the code
