+++ .devcontainer/devcontainer.json
@@ -34,9 +34,6 @@
       }
     }
   },
-  "features": {
-    "ghcr.io/gvatsal60/dev-container-features/sonarlint": "latest"
-  },
   "image": "mcr.microsoft.com/devcontainers/go:latest",
   "runArgs": [
     "--rm",
